requirements.txt
streamlit
google-generativeai
python-dotenv
seaborn
matplotlib
pandas
SpeechRecognition